fix: DateSelect: fix selection, remove placeholder workaround
Previously, react-select's <Select> component never actually knew which option was selected, because a *value* was being passing as `value`, instead of an option as intended; We merely kept track of the selected option/value within <DateSelect>'s parent's code. This is why the placeholder was always being shown, which in turn rendered the selected value.

Because we removed the placeholder (it isn't used anymore anyways, because we're passing in the select *option* now), we have to re-add the icon by wrapping react-select's <components.Control> component in a higher-order component that passes the icon in.